pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mothra"
version = "0.1.0"
description = "Master Agent Orchestration for Carbon Database Construction"
authors = [{name = "Mothra Team"}]
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.23",
    "pgvector>=0.2.4",
    "alembic>=1.13.0",
    "aiohttp>=3.9.1",
    "httpx>=0.25.2",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "playwright>=1.40.0",
    "pandas>=2.1.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "openai>=1.6.1",
    "tiktoken>=0.5.2",
    "tenacity>=8.2.3",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "prometheus-client>=0.19.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "tabula-py>=2.9.0",
    "openpyxl>=3.1.2",
    "xmltodict>=0.13.0",
    "pyyaml>=6.0.1",
    "fastapi>=0.115.0",
    "starlette>=0.40.0,<0.42.0",
    "uvicorn[standard]>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.7.1",
    "ipython>=8.19.0",
]

[tool.setuptools]
packages = ["mothra"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
